+++ CHANGELOG.md
@@ -17,6 +17,7 @@ have been replaced with `VideoLiveInfo`
 - Added `Serializable`
 - Changed `Video` to extend `Serializable`
 - All models are now const
+- Add String `code` enum `Language`
 
 ## 0.3.1